fix some warnings & autoformat

+++ src/lcd/lcd-controller.cpp
@@ -71,7 +71,7 @@ namespace gbaemu::lcd
         vcount = bitSet<uint16_t>(vcount, VCOUNT::CURRENT_SCANLINE_MASK, VCOUNT::CURRENT_SCANLINE_OFFSET, scanline.vCount);
         regsRef.VCOUNT = le(vcount);
 
-         /* update stat */
+        /* update stat */
         uint16_t stat = le(regsRef.DISPSTAT);
         stat = bitSet(stat, DISPSTAT::VBLANK_FLAG_MASK, DISPSTAT::VBLANK_FLAG_OFFSET, bmap<uint16_t>(scanline.vblanking));
         stat = bitSet(stat, DISPSTAT::HBLANK_FLAG_MASK, DISPSTAT::HBLANK_FLAG_OFFSET, bmap<uint16_t>(scanline.hblanking));
@@ -123,8 +123,8 @@ namespace gbaemu::lcd
         const color_t *src = frameBuffer.pixels();
         int32_t srcStride = frameBuffer.getWidth();
 
-        for (int32_t y = 0; y < SCREEN_HEIGHT; ++y) {
-            for (int32_t x = 0; x < SCREEN_WIDTH; ++x) {
+        for (int32_t y = 0; y < static_cast<int32_t>(SCREEN_HEIGHT); ++y) {
+            for (int32_t x = 0; x < static_cast<int32_t>(SCREEN_WIDTH); ++x) {
                 color_t color = src[y * srcStride + x];
 
                 for (int32_t sy = 0; sy < scale; ++sy) {
@@ -138,7 +138,7 @@ namespace gbaemu::lcd
         /* race conditions are acceptable here */
         *canDrawToScreen = true;
     }
- 
+
     std::string LCDController::getLayerStatusString() const
     {
         std::stringstream ss;
